chore: Move fakePubSub to utils and clean import tree
