+++ app/scripts/utils/fake-pub-sub.js
@@ -1,0 +1,8 @@
+const fakePubSub = {
+  __fake__: true,
+  publish: () => { },
+  subscribe: () => { },
+  unsubscribe: () => { },
+}
+
+export default fakePubSub;
