+++ test/tile-proxy.js
@@ -4,7 +4,7 @@ import { expect } from 'chai';
 import tileProxy, {
   tileDataToPixData,
 } from '../app/scripts/services/tile-proxy';
-import { fake as fakePubSub } from '../app/scripts/hocs/with-pub-sub';
+import fakePubSub from '../app/scripts/utils/fake-pub-sub';
 import {
   multivecTileData,
   multivecTileDataWithServerAggregation,
